requirements.txt
aioice==0.9.0
aiortc==1.9.0
av==12.3.0
boto3==1.35.10
botocore==1.35.10
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
dnspython==2.6.1
furl==2.1.3
gevent==24.2.1
google-crc32c==1.5.0
greenlet==3.0.3
idna==3.8
ifaddr==0.2.0
jmespath==1.0.1
Naked==0.1.32
orderedmultidict==1.0.1
pycparser==2.22
pycryptodome==3.20.0
pyee==12.0.0
pylibsrtp==0.10.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
setuptools==74.0.0
shellescape==3.8.1
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
websockets==13.0.1
zope.event==5.0
zope.interface==7.0.3
